requirements.txt
fastapi
uvicorn
python-dotenv
agno
openai
pypdf
gitpython
tiktoken
faiss-cpu
numpy
jinja2
